+++ src/components/Layout.tsx
@@ -1,7 +1,9 @@
 import { ReactNode, FC } from "react";
 import CssBaseline from "@mui/material/CssBaseline";
 import { Box, Container } from "@mui/material";
-import { Drawer, Header, Footer } from "~/components";
+import Drawer from "~/components/Drawer";
+import Header from "~/components/Header";
+import Footer from "~/components/Footer";
 import useUser from "~/lib/useUser";
 
 interface ILayout {
@@ -9,7 +11,7 @@ interface ILayout {
 }
 
 const Layout: FC<ILayout> = ({ children }) => {
-  const { user: session, mutateUser } = useUser();
+  const { user: session } = useUser();
   return (
     <Box
       sx={{
