+++ src/components/PaperWrapper.tsx
@@ -1,6 +1,6 @@
 import { Paper } from "@mui/material";
 import { FC, ReactNode } from "react";
-import { Title } from "~/components";
+import Title from "~/components/Title";
 
 interface IPaperWrapper {
   title?: string;
