+++ src/components/RegisterStepper/RegistrationInfo.tsx
