--- src/components/index.ts
@@ -1,39 +1,0 @@
-// --- layout
-export { default as HorizontalCenter } from "./helper/HorizontalCenter";
-export { default as Footer } from "./Footer";
-export { default as Header } from "./Header";
-export { default as Drawer } from "./Drawer";
-export { default as PaperWrapper } from "./PaperWrapper";
-
-// --- simple ui
-export { default as Title } from "./Title";
-export { default as InfoCard } from "./InfoCard";
-export { default as IsOnlineBadge } from "./IsOnlineBadge";
-export { default as StreamerCard } from "./StreamerCard";
-export { default as NumberInput } from "./NumberInput";
-export { default as LanguageSelector } from "./LanguageSelector";
-export { default as Subaddress } from "./Subaddress";
-export { default as SeedInput } from "./SeedInput";
-export { default as SeedOutput } from "./SeedOutput";
-export { default as HeroUnit } from "./HeroUnit";
-export { default as StreamerChip } from "./StreamerChip";
-
-// --- components
-export { default as Login } from "./Login";
-export { default as Transactions } from "./Transactions";
-
-// --- forms
-export { default as DonationSettingsForm } from "./settings/DonationSettings";
-export { default as WalletSettingsForm } from "./settings/WalletSettings";
-
-// --- invoice
-export { default as Invoice } from "./invoice/Invoice";
-
-// --- wallet
-export { default as TipxmrWallet } from "./wallet/TipxmrWallet";
-
-// --- donate
-export { default as DonationMask } from "./donate/DonationMask";
-
-// --- registration
-export { default as Register } from "./registration/Register";
