+++ src/pages/dashboard/index.tsx
@@ -1,6 +1,6 @@
 import { io } from "socket.io-client";
-import { IsOnlineBadge } from "~/components";
-import { useEffect, useState } from "react";
+import IsOnlineBadge from "~/components/IsOnlineBadge";
+import { useEffect } from "react";
 import Container from "@mui/material/Container";
 import Head from "next/head";
 import type { NextPage } from "next";
@@ -17,13 +17,13 @@ import {
   walletAtom,
 } from "~/store";
 import { createSyncProgressListener, open } from "~/lib/xmr";
-import { MoneroSubaddress } from "monero-javascript";
 
-const hash =
-  "b8185a25bbe3b4206e490558ab50b0567deca446d15282e92c5c66fde6693399".slice(
-    0,
-    11
-  );
+/* const hash =
+ *   "b8185a25bbe3b4206e490558ab50b0567deca446d15282e92c5c66fde6693399".slice(
+ *     0,
+ *     11
+ *   );
+ *  */
 
 function useSocket() {
   useEffect(() => {
