+++ src/pages/dashboard/settings.tsx
@@ -2,10 +2,10 @@ import { Grid } from "@mui/material";
 import { assoc } from "ramda";
 import { NextPage } from "next";
 import { FormEvent } from "react";
-import { DonationSettingsForm } from "~/components";
+import DonationSettingsForm from "~/components/DonationSettings";
 import fetchJson from "~/lib/fetchJson";
 import useUser from "~/lib/useUser";
-import { WalletSettingsForm } from "~/components";
+import WalletSettingsForm from "~/components/WalletSettings";
 import useDonationSettings from "~/hooks/useDonationSettings";
 import useWalletSettings from "~/hooks/useWalletSettings";
 import { useMutation, useQueryClient } from "react-query";
